fix: update task v3

+++ usecase/task/service.go
@@ -183,26 +183,28 @@ func (s Service) UpdateTask(data taskPayload.TaskUpdatePayload) error {
 	}
 
 	//Create task document
-	listTaskDocument := []*entity.TaskDocument{}
-	for _, documentPayload := range data.Documents {
-		document := &entity.TaskDocument{
-			File: documentPayload.File,
-			Name: documentPayload.Name,
-			TaskId: data.Id,
+	if data.Documents != nil {
+		listTaskDocument := []*entity.TaskDocument{}
+		for _, documentPayload := range data.Documents {
+			document := &entity.TaskDocument{
+				File: documentPayload.File,
+				Name: documentPayload.Name,
+				TaskId: data.Id,
+			}
+	
+			listTaskDocument = append(listTaskDocument, document)
 		}
-
-		listTaskDocument = append(listTaskDocument, document)
-	}
-
-	//Delete all task of document
-	err = s.taskDocumentRepo.DeleteAllTaskDocumentOfTask(data.Id)
-	if err != nil {
-		return err
-	}
-
-	err = s.taskDocumentRepo.CreateDocumentsForTask(listTaskDocument)
-	if err != nil {
-		return err
+	
+		//Delete all task of document
+		err = s.taskDocumentRepo.DeleteAllTaskDocumentOfTask(data.Id)
+		if err != nil {
+			return err
+		}
+	
+		err = s.taskDocumentRepo.CreateDocumentsForTask(listTaskDocument)
+		if err != nil {
+			return err
+		}
 	}
 	
 	return nil
